Convert Input to TS and fix resulting errors
