+++ src/components/inputs/InlineField.tsx
@@ -79,7 +79,6 @@ export default function InlineField({
       minHeight: inputHeight + paddingVertical * 2 + (android ? 2 : 0),
       paddingBottom: inputProps?.multiline ? (ios ? 15 : 7) : 0,
       paddingTop: inputProps?.multiline ? (android ? 11 : 15) : android ? (valueRef.current ? 16 : 11) : 0,
-      textAlignVertical: 'top',
       width: startsWith ? (ios ? 0.55 * width : 0.56 * width) : ios ? 0.6 * width : 0.61 * width,
     }),
     [textStyle, inputHeight, inputProps?.multiline, startsWith, width]
@@ -135,6 +134,7 @@ export default function InlineField({
             selectionColor={selectionColor}
             spellCheck={false}
             style={style}
+            textAlignVertical="top"
             value={value}
             {...inputProps}
             keyboardType={keyboardType}
